Change geolocation and regions to container component and render when needed
