+++ app/javascript/app/pages/root/root.js
@@ -2,11 +2,8 @@ import { PureComponent, createElement } from 'react';
 import Proptypes from 'prop-types';
 import { connect } from 'react-redux';
 import { withRouter } from 'react-router';
-import { actions as regionActions } from 'providers/regions-provider';
-import { actions as countriesActions } from 'providers/countries-provider';
+import { actions } from 'providers/countries-provider';
 import Component from './root-component';
-
-const actions = { ...regionActions, ...countriesActions };
 
 const mapStateToProps = state => ({
   countriesLoaded: state.countries.loaded
@@ -16,7 +13,6 @@ class Root extends PureComponent {
   constructor(props) {
     super(props);
     props.getCountries();
-    props.getRegions();
   }
 
   render() {
@@ -27,7 +23,6 @@ class Root extends PureComponent {
 }
 
 Root.propTypes = {
-  getRegions: Proptypes.func,
   getCountries: Proptypes.func,
   countriesLoaded: Proptypes.bool
 };
